Simplify props and extract NonMemberButtons
